- refactor regarding review suggestions
