+++ stake-pool/js/.gitignore
@@ -1,4 +1,3 @@
 dist
-package-lock.json
 .idea
 node_modules
