+++ stake-pool/js/.prettierrc
@@ -1,10 +1,8 @@
 {
   "arrowParens": "avoid",
   "bracketSpacing":false,
-  "jsxBracketSameLine": false,
   "singleQuote": true,
   "semi": false,
   "tabWidth": 2,
-  "trailingComma": "all",
-  "vueIndentScriptAndStyle": false
+  "trailingComma": "all"
 }
